+++ tencentcloud/cdn/v20180606/models.go
@@ -3483,6 +3483,9 @@ type PushUrlsCacheRequest struct {
 	// global：预热全球节点
 	// 不填充情况下，默认为 mainland， URL 中域名必须在对应区域启用了加速服务才能提交对应区域的预热任务
 	Area *string `json:"Area,omitempty" name:"Area"`
+
+	// 填写"middle"或不填充时预热至中间层节点
+	Layer *string `json:"Layer,omitempty" name:"Layer"`
 }
 
 func (r *PushUrlsCacheRequest) ToJsonString() string {