+++ tencentcloud/iotexplorer/v20190423/models.go
@@ -410,6 +410,14 @@ type DeleteDeviceResponse struct {
 	*tchttp.BaseResponse
 	Response *struct {
 
+		// 删除的结果代码
+	// 注意：此字段可能返回 null，表示取不到有效值。
+		ResultCode *string `json:"ResultCode,omitempty" name:"ResultCode"`
+
+		// 删除的结果信息
+	// 注意：此字段可能返回 null，表示取不到有效值。
+		ResultMessage *string `json:"ResultMessage,omitempty" name:"ResultMessage"`
+
 		// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 		RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
 	} `json:"Response"`